+++ common/src/main/java/foundry/veil/impl/client/render/shader/injection/util/Severity.java
@@ -1,0 +1,10 @@
+package foundry.veil.impl.client.render.shader.injection.util;
+
+import org.jetbrains.annotations.ApiStatus;
+
+@ApiStatus.Internal
+public enum Severity {
+    ERROR,
+    WARNING,
+    INFO
+}